% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bounds}
\alias{.bounds}
\title{Get bounding box from X/Y vectors}
\usage{
.bounds(x, y)
}
\arguments{
\item{x}{Vector of X coordinates}

\item{y}{Vector of Y coordinates}
}
\value{
Vector of class \code{bbox} with elements
\code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax}.
}
\description{
Get bounding box from X/Y vectors
}
\keyword{internal}
