% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_grid}
\alias{empty_grid}
\title{Get an empty matrix from X/Y coordinates}
\usage{
empty_grid(n)
}
\arguments{
\item{x}{Vector of X coordinates}

\item{y}{Vector of Y coordinates}
}
\value{
an \code{n} by \code{n} \code{matrix}.
}
\description{
Get an empty matrix from X/Y coordinates
}
\keyword{internal}
