% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapture.R
\name{rapture}
\alias{rapture}
\title{Create a raster capture \code{tmap}.}
\usage{
rapture(
  x,
  ...,
  basemap = TRUE,
  buffer = 10,
  map_service = "carto",
  map_type = "voyager_no_labels",
  downsample = FALSE,
  basemap_extra_args = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{An object coercable to a \code{stars} object.
Notably, this should be a \code{rapture_sparse} object.}

\item{...}{Passed to \code{tmap::tm_raster}.}

\item{buffer}{Amount to buffer tiles. See \code{sf::st_buffer}.}

\item{map_service}{character, a map service, either \code{"osm"}, \code{"carto"} or \code{"mapbox"}. Default is \code{"osm"}.}

\item{map_type}{character, a map type, e.g. \code{"streets"}. For a full list of available map types, see \code{\link[basemaps]{get_maptypes}}.}

\item{downsample}{Downsample raster/basemap if it's too big?}

\item{basemap_extra_args}{Named list of arguments passed
to \code{basemaps::basemap}.}

\item{verbose}{logical, if \code{TRUE}, messages and progress information are displayed on the console (default).}
}
\value{
A \code{tmap} object
}
\description{
Create a raster capture \code{tmap}.
}
