% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{sparse_grid}
\alias{sparse_grid}
\title{Get a sparse triples matrix from X/Y coordinates}
\usage{
sparse_grid(x, ...)
}
\arguments{
\item{x}{A vector of X coordinates, \code{data.frame},
\code{matrix}, or \code{rapture_dense} object.}

\item{...}{Unused.}

\item{y}{A vector of Y coordinates.}

\item{n}{Grid dimensions, \strong{must be a power of 2}.}

\item{coords}{Columns containing X/Y coordinates, respectively.}
}
\value{
Class of \code{dgTMatrix}
}
\description{
Get a sparse triples matrix from X/Y coordinates
}
