% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_as_stars.rapture_sparse}
\alias{st_as_stars.rapture_sparse}
\title{Coerce a \code{Matrix::sparseMatrix} to \code{stars}.}
\usage{
st_as_stars.rapture_sparse(.x, ...)
}
\arguments{
\item{x}{A matrix that inherits \code{Matrix::sparseMatrix}}
}
\description{
Coerce a \code{Matrix::sparseMatrix} to \code{stars}.
}
